requirements.txt
torch
torchvision
facenet-pytorch
matplotlib
numpy
Pillow
opencv-python
scikit-learn
